# PERF: impl Into for Duals 