Improvement for Dual into Dual2

For 100 variables:  Before: 22 us, After: 3 us
For 1 variable: Before 1.5us, After 0.5us